requirements-dev.txt
pytest~=6.0.2
eth-brownie
solc